Convert region AMIs to a list.
Per API guidelines, convert the map of regions to AMI to use, to a list of
structs which now also allow defining an optional master AMI to be used
if set. If none is set, masters will fall back to using the compute AMI
as we do today.

Logic to utilize when provisioning to follow.