Add SSHUser to cluster hardware info.
This is added to sit along-side the SSHKey.

Today this setting is largely dependent on the AMI and we require the
user to exist already.

Long term this setting will probably die, being replaced by global
cluster operator configuration for a user that we will ensure is created
on every host via the upcoming config DaemonSet work from kwoodson.
For now this gives us flexibility to adjust per cluster as ami's come
and go.