Complete functional cluster version installs.
Clusters will now install the version specified.

If there is no region defined on the version that matches the cluster,
machine set provisioning will be blocked and a condition added to the
cluster status.

YumRepositories and the previous specification of the AMI in the API are
both removed.

Upgrade (changing the cluster version for a cluster/machine set) is not
yet supported.