Replace openssl_certificate with custom certgen role.
The cert module in Ansible is a no-go on RHEL, and we don't want to rule
that out yet. Replace our older apiserver certgen script with a role we
can use for both apiserver certs, and individual cluster certs.

Generates a CA if one does not exist and uses it for all certs.

Each cluster gets it's own cert to prevent problems with the same cert
being uploaded anywhere in AWS.