Add cluster-api machine controller.
These are responsible for populating some portions of the machineSpec
providerConfig on machines.

- calculate the VMImage to use (uses cluster version and we expect
additional logic in future, which will only need to exist in the root
cluster)
- copy the cluster hardware spec for use in deletion (when cluster may
no longer exist)
- apply hardware defaults if applicable

These are only used in the root cluster, when we decide to send a
machineset down to the target cluster, we will update the spec
beforehand.

AWS actuator is updated to use the correct data going forward. (followup
PR will solve credentials to the actuator can actually create the master
machineset in the root cluster, presently this is still broken)