Fix bugs in AWS actuator around creation/deletion.
Fixes duplicate VMs being created when we're unable to set the instance
ID annotation because the machine was updated underneath us. Now looks
for instances that match our name which are pending or running.

Fix stuck machine records when the instance was deleted on AWS by
checking error condition and not considering it an error when the
instance ID no longer exists in AWS.

Added warnings for scenarios where we see multiple instances matching
our machine name.