Fix infra controller lookup of infra machineset size.
This was set up to look for a infra capi.MachineSet to get the size,
however we do not create this in the root cluster, so the infra job was
not spawning.

Instead it should use the co.Cluster.Spec.MachineSets just as we did
previously.

Integration tests required a rework as they did not configure the
machinesets on a co.ClusterSpec, used the co.ClusterMachineSet as the
new test input as this was easier to build capi.MachineSet off than the
reverse.

Applied a similar change for the components controller which was also
using the same logic for infra machine size lookup.