Fix clusterapi template master role and provider config.
In the absence of the now delete machine controller, manually setup the
provider config. This just allows testing until the flip to generate
everything off a ClusterDeployment is ready.

Fixes case on the Master role to match the upstream constant, which was
causing control plane install jobs not to trigger as it couldn't find a
master machine set.