Update machine status and allow for master deletion in root cluster.
Update machine status with core IP and hostname details as well as
instance ID.

AWS actuator now assumes multiple instances may exist matching it's name
and clusterid, and attempts to reconcile the situation if it occurs.

Fixed issues with deleting masters in the root cluster by specifying the
incluster flag.