stop including ansible vars with '-e'
using '-e' with ansible-playbook means that the variables included with the '-e' are immutable.

migrate the format of the vars file to be an ansible inventory file that applies to '[all]' hosts to avoid this immutability.

use INVENTORY_DIR when running the openshift-ansible container (instead of INVENTORY_FILE) to pull in the hosts and vars files in /ansible/inventory/.

fix handing of INVENTORY_DIR in openshift-ansible when the files are symlinks (it would copy the symlink instead of the contents of what the symlinks were pointing to). https://github.com/openshift/openshift-ansible/pull/9309 is the upstream fix.