label-based uninstaller
the label-based uninstaller tool will take a key=value as the basic filtering/matching mechanism to identify and remove AWS objects matching the tag/value.

there is special handling for untagged items (IAM objects in particular) where before deleting an object the tool will identify and remove any dependent objects that could otherwise not be found with tagging.

the general flow is that goroutines for each top-level-AWS-object type are launched and will spin (with backoff) and retry until each goroutine finds no more objects to delete.

the tool retries in the face of errors calling into the AWS API, but exceptional errors (like internal data manipulation and timeouts) cause the program to exit prematurely.